requirements.txt
torch>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0 